Fix: Only copy .service if not exists

+++ script/install.sh
@@ -9,5 +9,8 @@ sudo mkdir -p /usr/local/etc/latinaserver
 sudo cp $PROJECT/resources/openresty/stream/reverse_default.conf /usr/local/etc/latinaserver/
 sudo cp $PROJECT/config.json /usr/local/etc/latinaserver/
 sudo cp $PROJECT/latinaserver /usr/local/bin/ 
-sudo cp ./latinaserver.service /etc/systemd/system/
+
+if [ ! -f /etc/systemd/system/latinaserver.service ]; then
+    sudo cp ./latinaserver.service /etc/systemd/system/
+fi
 sudo systemctl daemon-reload
